refactor: move setLocaleCookie, getLocaleCookie to utils
